feat(builder): move part of config types to shared
