+++ .changeset/shiny-coats-mix.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/builder-shared': patch
+'@modern-js/builder-webpack-provider': patch
+---
+
+feat(builder): move part of config types to shared
+
+feat(builder): 将一部分公共的 config 类型定义下沉到 shared 中
