+++ packages/builder/builder-shared/src/types/config/output.ts
@@ -1,0 +1,79 @@
+export type DistPathConfig = {
+  root?: string;
+  js?: string;
+  css?: string;
+  svg?: string;
+  font?: string;
+  html?: string;
+  image?: string;
+  media?: string;
+  server?: string;
+};
+
+export type FilenameConfig = {
+  js?: string;
+  css?: string;
+  svg?: string;
+  font?: string;
+  image?: string;
+  media?: string;
+};
+
+export type DataUriLimit = {
+  svg?: number;
+  font?: number;
+  image?: number;
+  media?: number;
+};
+
+export type Charset = 'ascii' | 'utf8';
+
+export type SvgDefaultExport = 'component' | 'url';
+
+export type LegalComments = 'none' | 'inline' | 'linked';
+
+export type NormalizedDataUriLimit = Required<DataUriLimit>;
+
+export type Polyfill = 'usage' | 'entry' | 'ua' | 'off';
+
+export interface SharedOutputConfig {
+  distPath?: DistPathConfig;
+  filename?: FilenameConfig;
+  charset?: Charset;
+  polyfill?: Polyfill;
+  assetPrefix?: string;
+  dataUriLimit?: number | DataUriLimit;
+  legalComments?: LegalComments;
+  cleanDistPath?: boolean;
+  disableMinimize?: boolean;
+  disableSourceMap?: boolean;
+  disableFilenameHash?: boolean;
+  disableInlineRuntimeChunk?: boolean;
+  enableAssetManifest?: boolean;
+  enableAssetFallback?: boolean;
+  enableLatestDecorators?: boolean;
+  enableCssModuleTSDeclaration?: boolean;
+  enableInlineScripts?: boolean;
+  enableInlineStyles?: boolean;
+  overrideBrowserslist?: string[];
+  svgDefaultExport?: SvgDefaultExport;
+}
+
+export interface NormalizedSharedOutputConfig extends SharedOutputConfig {
+  filename: FilenameConfig;
+  distPath: DistPathConfig;
+  polyfill: Polyfill;
+  dataUriLimit: NormalizedDataUriLimit;
+  cleanDistPath: boolean;
+  disableMinimize: boolean;
+  disableSourceMap: boolean;
+  disableFilenameHash: boolean;
+  disableInlineRuntimeChunk: boolean;
+  enableAssetManifest: boolean;
+  enableAssetFallback: boolean;
+  enableLatestDecorators: boolean;
+  enableCssModuleTSDeclaration: boolean;
+  enableInlineScripts: boolean;
+  enableInlineStyles: boolean;
+  svgDefaultExport: SvgDefaultExport;
+}
