+++ packages/builder/builder-shared/src/types/config/security.ts
@@ -1,0 +1,9 @@
+export type SriOptions = {
+  hashFuncNames?: [string, ...string[]];
+  enabled?: 'auto' | true | false;
+  hashLoading?: 'eager' | 'lazy';
+};
+
+export interface SharedSecurityConfig {
+  sri?: SriOptions | boolean;
+}
