+++ packages/builder/builder-shared/src/types/index.ts
@@ -3,3 +3,4 @@ export * from './context';
 export * from './utils';
 export * from './plugin';
 export * from './provider';
+export * from './config';
