+++ packages/builder/builder-webpack-provider/src/types/config/source.ts
@@ -1,30 +1,8 @@
-import type { ChainedConfig, JSONValue } from '@modern-js/builder-shared';
-import type { WebpackAlias } from '../thirdParty';
-import type * as webpack from 'webpack';
+import type {
+  SharedSourceConfig,
+  NormalizedSharedSourceConfig,
+} from '@modern-js/builder-shared';
 
-export type ModuleScopes = Array<string | RegExp>;
-export type CodeValue = webpack.DefinePlugin['definitions'][string];
+export type SourceConfig = SharedSourceConfig;
 
-export interface SourceConfig {
-  include?: (string | RegExp)[];
-  alias?: ChainedConfig<WebpackAlias>;
-  preEntry?: string | string[];
-  globalVars?: Record<string, JSONValue>;
-  define?: Record<string, CodeValue>;
-  moduleScopes?: ChainedConfig<ModuleScopes>;
-  compileJsDataURI?: boolean;
-  resolveExtensionPrefix?: string;
-  resolveMainFields?: (string[] | string)[];
-}
-
-export interface SourceFinalConfig {
-  preEntry?: string | string[];
-  resolveExtensionPrefix?: string;
-}
-
-export interface NormalizedSourceConfig extends SourceConfig {
-  preEntry: string[];
-  globalVars: Record<string, JSONValue>;
-  define: Record<string, CodeValue>;
-  compileJsDataURI: boolean;
-}
+export type NormalizedSourceConfig = NormalizedSharedSourceConfig;
